+++ app/Http/Controllers/Admin/Project/TaskReminderController.php
@@ -11,11 +11,15 @@ class TaskReminderController extends Controller
 {
   public function index($project, Task $task)
   {
+    abort_if(!$task->project->isMine(), 403);
+
     return $this->sendRes('success', ['view_data' => view('admin.pages.projects.tasks.reminders-index', compact('task'))->render()]);
   }
 
   public function store($project, Task $task, Request $request)
   {
+    abort_if(!$task->project->isMine(), 403);
+
     $request->validate([
       'recipient_id' => 'required|exists:admins,id',
       'description' => 'required',
@@ -33,7 +37,10 @@ class TaskReminderController extends Controller
 
   public function destroy($project, $task, TaskReminder $reminder)
   {
+    abort_if(!$reminder->task->project->isMine(), 403);
+
     $reminder->delete();
+
     return $this->sendRes('Reminder deleted successfully', []);
   }
 }
